@@ -8,23 +8,24 @@ issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
+  collection: ^1.15.0-nullsafety.4
   flutter:
     sdk: flutter
-  meta: ^1.2.4
+  meta: ^1.3.0
   rxdart: ^0.26.0
-  stream_chat: ^1.5.0
+  stream_chat: ^1.5.2
 
 dependency_overrides:
   stream_chat:
     path: ../stream_chat
 
 dev_dependencies:
-  fake_async: ^1.1.0
+  fake_async: ^1.2.0
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.3
+  mocktail: ^0.1.1
 
