@@ -13,10 +13,9 @@ import 'package:stream_chat_flutter_core/src/stream_chat_core.dart';
 class MessageSearchBloc extends StatefulWidget {
   /// Instantiate a new MessageSearchBloc
   const MessageSearchBloc({
-    Key key,
-    @required this.child,
-  })  : assert(child != null, 'Parameter child should not be null.'),
-        super(key: key);
+    Key? key,
+    required this.child,
+  }) : super(key: key);
 
   /// The widget child
   final Widget child;
@@ -26,7 +25,7 @@ class MessageSearchBloc extends StatefulWidget {
 
   /// Use this method to get the current [MessageSearchBlocState] instance
   static MessageSearchBlocState of(BuildContext context) {
-    MessageSearchBlocState state;
+    MessageSearchBlocState? state;
 
     state = context.findAncestorStateOfType<MessageSearchBlocState>();
 
@@ -42,7 +41,7 @@ class MessageSearchBloc extends StatefulWidget {
 class MessageSearchBlocState extends State<MessageSearchBloc>
     with AutomaticKeepAliveClientMixin {
   /// The current messages list
-  List<GetMessageResponse> get messageResponses => _messageResponses.value;
+  List<GetMessageResponse>? get messageResponses => _messageResponses.value;
 
   /// The current messages list as a stream
   Stream<List<GetMessageResponse>> get messagesStream =>
@@ -59,11 +58,11 @@ class MessageSearchBlocState extends State<MessageSearchBloc>
   /// Calls [StreamChatClient.search] updating
   /// [messagesStream] and [queryMessagesLoading] stream
   Future<void> search({
-    Map<String, dynamic> filter,
-    Map<String, dynamic> messageFilter,
-    List<SortOption> sort,
-    String query,
-    PaginationParams pagination,
+    required Map<String, dynamic> filter,
+    Map<String, dynamic>? messageFilter,
+    List<SortOption>? sort,
+    String? query,
+    PaginationParams? pagination,
   }) async {
     final client = StreamChatCore.of(context).client;
 
@@ -73,9 +72,7 @@ class MessageSearchBlocState extends State<MessageSearchBloc>
       _queryMessagesLoadingController.add(true);
     }
     try {
-      final clear = pagination == null ||
-          pagination.offset == null ||
-          pagination.offset == 0;
+      final clear = pagination == null || pagination.offset == 0;
 
       final oldMessages = List<GetMessageResponse>.from(messageResponses ?? []);
 
@@ -93,7 +90,8 @@ class MessageSearchBlocState extends State<MessageSearchBloc>
         final temp = oldMessages + messages.results;
         _messageResponses.add(temp);
       }
-      if (_messageResponses.hasValue && _queryMessagesLoadingController.value) {
+      if (_messageResponses.hasValue &&
+          _queryMessagesLoadingController.value!) {
         _queryMessagesLoadingController.add(false);
       }
     } catch (e, stk) {
