@@ -42,31 +42,27 @@ class MessageSearchListCore extends StatefulWidget {
   /// * [loadingBuilder]
   /// * [childBuilder]
   const MessageSearchListCore({
-    Key key,
-    @required this.emptyBuilder,
-    @required this.errorBuilder,
-    @required this.loadingBuilder,
-    @required this.childBuilder,
+    Key? key,
+    required this.emptyBuilder,
+    required this.errorBuilder,
+    required this.loadingBuilder,
+    required this.childBuilder,
+    required this.filters,
     this.messageQuery,
-    this.filters,
     this.sortOptions,
     this.paginationParams,
     this.messageFilters,
     this.messageSearchListController,
-  })  : assert(emptyBuilder != null, 'emptyBuilder should not be null'),
-        assert(errorBuilder != null, 'errorBuilder should not be null'),
-        assert(loadingBuilder != null, 'loadingBuilder should not be null'),
-        assert(childBuilder != null, 'childBuilder should not be null'),
-        super(key: key);
+  }) : super(key: key);
 
   /// A [MessageSearchListController] allows reloading and pagination.
   /// Use [MessageSearchListController.loadData] and
   /// [MessageSearchListController.paginateData] respectively for reloading and
   /// pagination.
-  final MessageSearchListController messageSearchListController;
+  final MessageSearchListController? messageSearchListController;
 
   /// Message String to search on
-  final String messageQuery;
+  final String? messageQuery;
 
   /// The query filters to use.
   /// You can query on any of the custom fields you've defined on the [Channel].
@@ -78,18 +74,18 @@ class MessageSearchListCore extends StatefulWidget {
   /// provided.
   /// You can sort based on last_updated, last_message_at, updated_at, created_
   /// at or member_count. Direction can be ascending or descending.
-  final List<SortOption> sortOptions;
+  final List<SortOption>? sortOptions;
 
   /// Pagination parameters
   /// limit: the number of users to return (max is 30)
   /// offset: the offset (max is 1000)
   /// message_limit: how many messages should be included to each channel
-  final PaginationParams paginationParams;
+  final PaginationParams? paginationParams;
 
   /// The message query filters to use.
   /// You can query on any of the custom fields you've defined on the [Channel].
   /// You can also filter other built-in channel fields.
-  final Map<String, dynamic> messageFilters;
+  final Map<String, dynamic>? messageFilters;
 
   /// The builder that is used when the search messages are fetched
   final Widget Function(List<GetMessageResponse>) childBuilder;
@@ -114,8 +110,8 @@ class MessageSearchListCoreState extends State<MessageSearchListCore> {
     super.didChangeDependencies();
     loadData();
     if (widget.messageSearchListController != null) {
-      widget.messageSearchListController.loadData = loadData;
-      widget.messageSearchListController.paginateData = paginateData;
+      widget.messageSearchListController!.loadData = loadData;
+      widget.messageSearchListController!.paginateData = paginateData;
     }
   }
 
@@ -130,16 +126,16 @@ class MessageSearchListCoreState extends State<MessageSearchListCore> {
         stream: messageSearchBloc.messagesStream,
         builder: (context, snapshot) {
           if (snapshot.hasError) {
-            return widget.errorBuilder(context, snapshot.error);
+            return widget.errorBuilder(context, snapshot.error!);
           }
           if (!snapshot.hasData) {
             return widget.loadingBuilder(context);
           }
-          final items = snapshot.data;
+          final items = snapshot.data!;
           if (items.isEmpty) {
             return widget.emptyBuilder(context);
           }
-          return widget.childBuilder(snapshot.data);
+          return widget.childBuilder(items);
         },
       );
 
@@ -161,7 +157,7 @@ class MessageSearchListCoreState extends State<MessageSearchListCore> {
     return messageSearchBloc.search(
       filter: widget.filters,
       sort: widget.sortOptions,
-      pagination: widget.paginationParams.copyWith(
+      pagination: widget.paginationParams!.copyWith(
         offset: messageSearchBloc.messageResponses?.length ?? 0,
       ),
       query: widget.messageQuery,
@@ -172,13 +168,13 @@ class MessageSearchListCoreState extends State<MessageSearchListCore> {
   @override
   void didUpdateWidget(MessageSearchListCore oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (widget.filters?.toString() != oldWidget.filters?.toString() ||
+    if (widget.filters.toString() != oldWidget.filters.toString() ||
         jsonEncode(widget.sortOptions) != jsonEncode(oldWidget.sortOptions) ||
         widget.messageQuery?.toString() != oldWidget.messageQuery?.toString() ||
         widget.messageFilters?.toString() !=
             oldWidget.messageFilters?.toString() ||
-        widget.paginationParams?.toJson()?.toString() !=
-            oldWidget.paginationParams?.toJson()?.toString()) {
+        widget.paginationParams?.toJson().toString() !=
+            oldWidget.paginationParams?.toJson().toString()) {
       loadData();
     }
   }
@@ -187,8 +183,8 @@ class MessageSearchListCoreState extends State<MessageSearchListCore> {
 /// Controller used for paginating data in [ChannelListView]
 class MessageSearchListController {
   /// Call this function to reload data
-  AsyncCallback loadData;
+  AsyncCallback? loadData;
 
   /// Call this function to load further data
-  AsyncCallback paginateData;
+  AsyncCallback? paginateData;
 }
