@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:rxdart/rxdart.dart';
@@ -21,16 +22,14 @@ class StreamChannel extends StatefulWidget {
   /// Creates a new instance of [StreamChannel]. Both [child] and [client] must
   /// be supplied and not null.
   const StreamChannel({
-    Key key,
-    @required this.child,
-    @required this.channel,
+    Key? key,
+    required this.child,
+    required this.channel,
     this.showLoading = true,
     this.initialMessageId,
-  })  : assert(child != null, 'Child should not be null'),
-        assert(channel != null, 'Channel should not be null'),
-        super(key: key);
+  }) : super(key: key);
 
-  // ignore: public_member_api_docs
+  /// The child of the widget
   final Widget child;
 
   /// [channel] specifies the channel with which child should be wrapped
@@ -40,11 +39,11 @@ class StreamChannel extends StatefulWidget {
   final bool showLoading;
 
   /// If passed the channel will load from this particular message.
-  final String initialMessageId;
+  final String? initialMessageId;
 
   /// Use this method to get the current [StreamChannelState] instance
   static StreamChannelState of(BuildContext context) {
-    StreamChannelState streamChannelState;
+    StreamChannelState? streamChannelState;
 
     streamChannelState = context.findAncestorStateOfType<StreamChannelState>();
 
@@ -67,11 +66,11 @@ class StreamChannelState extends State<StreamChannel> {
   Channel get channel => widget.channel;
 
   /// InitialMessageId
-  String get initialMessageId => widget.initialMessageId;
+  String? get initialMessageId => widget.initialMessageId;
 
   /// Current channel state stream
-  Stream<ChannelState> get channelStateStream =>
-      widget.channel.state.channelStateStream;
+  Stream<ChannelState>? get channelStateStream =>
+      widget.channel.state?.channelStateStream;
 
   final _queryTopMessagesController = BehaviorSubject.seeded(false);
   final _queryBottomMessagesController = BehaviorSubject.seeded(false);
@@ -89,16 +88,18 @@ class StreamChannelState extends State<StreamChannel> {
     int limit = 20,
     bool preferOffline = false,
   }) async {
-    if (_topPaginationEnded || _queryTopMessagesController?.value == true) {
+    if (_topPaginationEnded ||
+        _queryTopMessagesController.value == true ||
+        channel.state == null) {
       return;
     }
     _queryTopMessagesController.add(true);
 
-    if (channel.state.messages.isEmpty) {
+    if (channel.state!.messages.isEmpty) {
       return _queryTopMessagesController.add(false);
     }
 
-    final oldestMessage = channel.state.messages.first;
+    final oldestMessage = channel.state!.messages.first;
 
     try {
       final state = await queryBeforeMessage(
@@ -120,15 +121,16 @@ class StreamChannelState extends State<StreamChannel> {
     bool preferOffline = false,
   }) async {
     if (_bottomPaginationEnded ||
-        _queryBottomMessagesController?.value == true ||
-        channel?.state?.isUpToDate == true) return;
+        _queryBottomMessagesController.value == true ||
+        channel.state == null ||
+        channel.state!.isUpToDate == true) return;
     _queryBottomMessagesController.add(true);
 
-    if (channel.state.messages.isEmpty) {
+    if (channel.state!.messages.isEmpty) {
       return _queryBottomMessagesController.add(false);
     }
 
-    final recentMessage = channel.state.messages.last;
+    final recentMessage = channel.state!.messages.last;
 
     try {
       final state = await queryAfterMessage(
@@ -146,7 +148,7 @@ class StreamChannelState extends State<StreamChannel> {
   }
 
   /// Calls [channel.query] updating [queryMessage] stream
-  Future<void> queryMessages({QueryDirection direction = QueryDirection.top}) {
+  Future<void> queryMessages({QueryDirection? direction = QueryDirection.top}) {
     if (direction == QueryDirection.top) return _queryTopMessages();
     return _queryBottomMessages();
   }
@@ -157,12 +159,14 @@ class StreamChannelState extends State<StreamChannel> {
     int limit = 50,
     bool preferOffline = false,
   }) async {
-    if (_topPaginationEnded || _queryTopMessagesController.value) return;
+    if (_topPaginationEnded ||
+        _queryTopMessagesController.value! ||
+        channel.state == null) return;
     _queryTopMessagesController.add(true);
 
-    Message message;
-    if (channel.state.threads.containsKey(parentId)) {
-      final thread = channel.state.threads[parentId];
+    late Message message;
+    if (channel.state!.threads.containsKey(parentId)) {
+      final thread = channel.state!.threads[parentId]!;
       if (thread.isNotEmpty) {
         message = thread.first;
       }
@@ -172,7 +176,7 @@ class StreamChannelState extends State<StreamChannel> {
       final response = await channel.getReplies(
         parentId,
         PaginationParams(
-          lessThan: message?.id,
+          lessThan: message.id,
           limit: limit,
         ),
         preferOffline: preferOffline,
@@ -188,21 +192,26 @@ class StreamChannelState extends State<StreamChannel> {
 
   /// Query the channel members and watchers
   Future<void> queryMembersAndWatchers() async {
-    await widget.channel.query(
-      membersPagination: PaginationParams(
-        offset: channel.state.members?.length,
-        limit: 100,
-      ),
-      watchersPagination: PaginationParams(
-        offset: channel.state.watchers?.length,
-        limit: 100,
-      ),
-    );
+    final _members = channel.state?.members;
+    if (_members != null) {
+      await widget.channel.query(
+        membersPagination: PaginationParams(
+          offset: _members.length,
+          limit: 100,
+        ),
+        watchersPagination: PaginationParams(
+          offset: _members.length,
+          limit: 100,
+        ),
+      );
+    } else {
+      return;
+    }
   }
 
   /// Loads channel at specific message
   Future<void> loadChannelAtMessage(
-    String messageId, {
+    String? messageId, {
     int before = 20,
     int after = 20,
     bool preferOffline = false,
@@ -214,15 +223,15 @@ class StreamChannelState extends State<StreamChannel> {
         preferOffline: preferOffline,
       );
 
-  Future<void> _queryAtMessage({
-    String messageId,
+  Future<List<ChannelState>> _queryAtMessage({
+    String? messageId,
     int before = 20,
     int after = 20,
     bool preferOffline = false,
   }) async {
-    if (channel.state == null) return;
-    channel.state.isUpToDate = false;
-    channel.state.truncate();
+    if (channel.state == null) return [];
+    channel.state!.isUpToDate = false;
+    channel.state!.truncate();
 
     if (messageId == null) {
       await channel.query(
@@ -231,8 +240,8 @@ class StreamChannelState extends State<StreamChannel> {
         ),
         preferOffline: preferOffline,
       );
-      channel.state.isUpToDate = true;
-      return;
+      channel.state!.isUpToDate = true;
+      return [];
     }
 
     return Future.wait([
@@ -277,16 +286,15 @@ class StreamChannelState extends State<StreamChannel> {
       preferOffline: preferOffline,
     );
     if (state.messages.isEmpty || state.messages.length < limit) {
-      channel.state.isUpToDate = true;
+      channel.state?.isUpToDate = true;
     }
     return state;
   }
 
   ///
   Future<Message> getMessage(String messageId) async {
-    var message = channel.state.messages.firstWhere(
+    var message = channel.state?.messages.firstWhereOrNull(
       (it) => it.id == messageId,
-      orElse: () => null,
     );
     if (message == null) {
       final response = await channel.getMessagesById([messageId]);
@@ -298,7 +306,7 @@ class StreamChannelState extends State<StreamChannel> {
   /// Reloads the channel with latest message
   Future<void> reloadChannel() => _queryAtMessage(before: 30);
 
-  List<Future<bool>> _futures;
+  late List<Future<bool>> _futures;
 
   Future<bool> get _loadChannelAtMessage async {
     try {
@@ -358,9 +366,9 @@ class StreamChannelState extends State<StreamChannel> {
           }
           return Center(child: Text(message));
         }
-        final initialized = snapshot.data[0];
+        final initialized = snapshot.data![0];
         // ignore: avoid_bool_literals_in_conditional_expressions
-        final dataLoaded = initialMessageId == null ? true : snapshot.data[1];
+        final dataLoaded = initialMessageId == null ? true : snapshot.data![1];
         if (widget.showLoading && (!initialized || !dataLoaded)) {
           return const Center(
             child: CircularProgressIndicator(),
