@@ -14,13 +14,9 @@ class UsersBloc extends StatefulWidget {
   /// Instantiate a new [UsersBloc]. The parameter [child] must be supplied and
   /// not null.
   const UsersBloc({
-    @required this.child,
-    Key key,
-  })  : assert(
-            child != null,
-            'When constructing a UsersBloc, the parameter '
-            'child should not be null.'),
-        super(key: key);
+    required this.child,
+    Key? key,
+  }) : super(key: key);
 
   /// The widget child
   final Widget child;
@@ -30,7 +26,7 @@ class UsersBloc extends StatefulWidget {
 
   /// Use this method to get the current [UsersBlocState] instance
   static UsersBlocState of(BuildContext context) {
-    UsersBlocState state;
+    UsersBlocState? state;
 
     state = context.findAncestorStateOfType<UsersBlocState>();
 
@@ -46,7 +42,7 @@ class UsersBloc extends StatefulWidget {
 class UsersBlocState extends State<UsersBloc>
     with AutomaticKeepAliveClientMixin {
   /// The current users list
-  List<User> get users => _usersController.value;
+  List<User>? get users => _usersController.value;
 
   /// The current users list as a stream
   Stream<List<User>> get usersStream => _usersController.stream;
@@ -62,10 +58,10 @@ class UsersBlocState extends State<UsersBloc>
   /// online/offline.
   /// [API Reference](https://getstream.io/chat/docs/flutter-dart/query_users/?language=dart)
   Future<void> queryUsers({
-    Map<String, dynamic> filter,
-    List<SortOption> sort,
-    Map<String, dynamic> options,
-    PaginationParams pagination,
+    Map<String, dynamic>? filter,
+    List<SortOption>? sort,
+    Map<String, dynamic>? options,
+    PaginationParams? pagination,
   }) async {
     final client = StreamChatCore.of(context).client;
 
@@ -76,9 +72,7 @@ class UsersBlocState extends State<UsersBloc>
     }
 
     try {
-      final clear = pagination == null ||
-          pagination.offset == null ||
-          pagination.offset == 0;
+      final clear = pagination == null || pagination.offset == 0;
 
       final oldUsers = List<User>.from(users ?? []);
 
@@ -95,7 +89,7 @@ class UsersBlocState extends State<UsersBloc>
         final temp = oldUsers + usersResponse.users;
         _usersController.add(temp);
       }
-      if (_usersController.hasValue && _queryUsersLoadingController.value) {
+      if (_usersController.hasValue && _queryUsersLoadingController.value!) {
         _queryUsersLoadingController.add(false);
       }
     } catch (e, stk) {
