@@ -1,4 +1,3 @@
-import 'package:meta/meta.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
@@ -7,15 +6,15 @@ Matcher isSameMessageResponseAs(GetMessageResponse targetResponse) =>
 
 class _IsSameMessageResponseAs extends Matcher {
   const _IsSameMessageResponseAs({
-    @required this.targetResponse,
-  }) : assert(targetResponse != null, '');
+    required this.targetResponse,
+  });
 
   final GetMessageResponse targetResponse;
 
   @override
   bool matches(covariant GetMessageResponse response, Map matchState) =>
       response.message.id == targetResponse.message.id &&
-      response.channel.cid == targetResponse.channel.cid;
+      response.channel?.cid == targetResponse.channel?.cid;
 
   @override
   Description describe(Description description) =>
@@ -28,15 +27,15 @@ Matcher isSameMessageResponseListAs(
 
 class _IsSameMessageResponseListAs extends Matcher {
   const _IsSameMessageResponseListAs({
-    @required this.targetResponseList,
-  }) : assert(targetResponseList != null, '');
+    required this.targetResponseList,
+  });
 
   final List<GetMessageResponse> targetResponseList;
 
   @override
   bool matches(
       covariant List<GetMessageResponse> responseList, Map matchState) {
-    bool matches = true;
+    var matches = true;
     for (var i = 0; i < responseList.length; i++) {
       final response = responseList[i];
       final targetResponse = targetResponseList[i];
