@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
-import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat_flutter_core/src/message_search_list_core.dart';
+import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 import 'mocks.dart';
 
@@ -10,74 +10,21 @@ void main() {
   List<GetMessageResponse> _generateMessages({
     int count = 3,
     int offset = 0,
-  }) {
-    return List.generate(
-      count,
-      (index) {
-        index = index + offset;
-        return GetMessageResponse()
-          ..message = Message(
-            id: 'testId$index',
-            text: 'testTextData$index',
-          )
-          ..channel = ChannelModel(
-            cid: 'testCid',
-          );
-      },
-    );
-  }
-
-  test(
-    'should throw assertion error in case childBuilder is null',
-    () {
-      final messageSearchListCore = () => MessageSearchListCore(
-            childBuilder: null,
-            loadingBuilder: (BuildContext context) => Offstage(),
-            emptyBuilder: (BuildContext context) => Offstage(),
-            errorBuilder: (BuildContext context, Object error) => Offstage(),
-          );
-      expect(messageSearchListCore, throwsA(isA<AssertionError>()));
-    },
-  );
-
-  test(
-    'should throw assertion error in case loadingBuilder is null',
-    () {
-      final messageSearchListCore = () => MessageSearchListCore(
-            childBuilder: (List<GetMessageResponse> messages) => Offstage(),
-            loadingBuilder: null,
-            emptyBuilder: (BuildContext context) => Offstage(),
-            errorBuilder: (BuildContext context, Object error) => Offstage(),
-          );
-      expect(messageSearchListCore, throwsA(isA<AssertionError>()));
-    },
-  );
-
-  test(
-    'should throw assertion error in case emptyBuilder is null',
-    () {
-      final messageSearchListCore = () => MessageSearchListCore(
-            childBuilder: (List<GetMessageResponse> messages) => Offstage(),
-            loadingBuilder: (BuildContext context) => Offstage(),
-            emptyBuilder: null,
-            errorBuilder: (BuildContext context, Object error) => Offstage(),
-          );
-      expect(messageSearchListCore, throwsA(isA<AssertionError>()));
-    },
-  );
-
-  test(
-    'should throw assertion error in case errorBuilder is null',
-    () {
-      final messageSearchListCore = () => MessageSearchListCore(
-            childBuilder: (List<GetMessageResponse> messages) => Offstage(),
-            loadingBuilder: (BuildContext context) => Offstage(),
-            emptyBuilder: (BuildContext context) => Offstage(),
-            errorBuilder: null,
-          );
-      expect(messageSearchListCore, throwsA(isA<AssertionError>()));
-    },
-  );
+  }) =>
+      List.generate(
+        count,
+        (index) {
+          index = index + offset;
+          return GetMessageResponse()
+            ..message = Message(
+              id: 'testId$index',
+              text: 'testTextData$index',
+            )
+            ..channel = ChannelModel(
+              cid: 'test:Cid',
+            );
+        },
+      );
 
   testWidgets(
     'should throw if MessageSearchListCore is used where MessageSearchBloc '
@@ -86,10 +33,11 @@ void main() {
       const messageSearchListCoreKey = Key('messageSearchListCore');
       final messageSearchListCore = MessageSearchListCore(
         key: messageSearchListCoreKey,
-        childBuilder: (List<GetMessageResponse> messages) => Offstage(),
-        loadingBuilder: (BuildContext context) => Offstage(),
-        emptyBuilder: (BuildContext context) => Offstage(),
-        errorBuilder: (BuildContext context, Object error) => Offstage(),
+        childBuilder: (List<GetMessageResponse>? messages) => const Offstage(),
+        loadingBuilder: (BuildContext context) => const Offstage(),
+        emptyBuilder: (BuildContext context) => const Offstage(),
+        errorBuilder: (BuildContext context, Object? error) => const Offstage(),
+        filters: const {},
       );
 
       await tester.pumpWidget(messageSearchListCore);
@@ -109,7 +57,8 @@ void main() {
         childBuilder: (List<GetMessageResponse> messages) => Offstage(),
         loadingBuilder: (BuildContext context) => Offstage(),
         emptyBuilder: (BuildContext context) => Offstage(),
-        errorBuilder: (BuildContext context, Object error) => Offstage(),
+        errorBuilder: (BuildContext context, Object? error) => Offstage(),
+        filters: {},
       );
 
       final mockClient = MockClient();
@@ -140,6 +89,7 @@ void main() {
         emptyBuilder: (BuildContext context) => Offstage(),
         errorBuilder: (BuildContext context, Object error) => Offstage(),
         messageSearchListController: controller,
+        filters: {},
       );
 
       expect(controller.loadData, isNull);
@@ -175,18 +125,19 @@ void main() {
         errorBuilder: (BuildContext context, Object error) => Offstage(
           key: errorWidgetKey,
         ),
+        filters: {},
       );
 
       final mockClient = MockClient();
 
       const error = 'Error! Error! Error!';
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenThrow(error);
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenThrow(error);
 
       await tester.pumpWidget(
         StreamChatCore(
@@ -201,13 +152,13 @@ void main() {
 
       expect(find.byKey(errorWidgetKey), findsOneWidget);
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
     },
   );
 
@@ -223,18 +174,19 @@ void main() {
         loadingBuilder: (BuildContext context) => Offstage(),
         emptyBuilder: (BuildContext context) => Offstage(key: emptyWidgetKey),
         errorBuilder: (BuildContext context, Object error) => Offstage(),
+        filters: {},
       );
 
       final mockClient = MockClient();
 
       final messageResponseList = <GetMessageResponse>[];
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
@@ -251,13 +203,13 @@ void main() {
 
       expect(find.byKey(emptyWidgetKey), findsOneWidget);
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
     },
   );
 
@@ -274,18 +226,19 @@ void main() {
         loadingBuilder: (BuildContext context) => Offstage(),
         emptyBuilder: (BuildContext context) => Offstage(),
         errorBuilder: (BuildContext context, Object error) => Offstage(),
+        filters: {},
       );
 
       final mockClient = MockClient();
 
       final messageResponseList = _generateMessages();
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
@@ -302,13 +255,13 @@ void main() {
 
       expect(find.byKey(childWidgetKey), findsOneWidget);
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
     },
   );
 
@@ -324,25 +277,26 @@ void main() {
         childBuilder: (List<GetMessageResponse> messages) => Container(
           key: childWidgetKey,
           child: Text(
-            messages.map((e) => '${e.channel.cid}-${e.message.id}').join(','),
+            messages.map((e) => '${e.channel?.cid}-${e.message.id}').join(','),
           ),
         ),
         loadingBuilder: (BuildContext context) => Offstage(),
         emptyBuilder: (BuildContext context) => Offstage(),
         errorBuilder: (BuildContext context, Object error) => Offstage(),
         paginationParams: pagination,
+        filters: {},
       );
 
       final mockClient = MockClient();
 
       final messageResponseList = _generateMessages();
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
@@ -364,19 +318,19 @@ void main() {
       expect(
         find.text(
           messageResponseList
-              .map((e) => '${e.channel.cid}-${e.message.id}')
+              .map((e) => '${e.channel?.cid}-${e.message.id}')
               .join(','),
         ),
         findsOneWidget,
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).called(1);
 
       final messageSearchListCoreState =
           tester.state<MessageSearchListCoreState>(
@@ -386,13 +340,13 @@ void main() {
       final offset = messageResponseList.length;
       final paginatedMessageResponseList = _generateMessages(offset: offset);
       final updatedPagination = pagination.copyWith(offset: offset);
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: updatedPagination,
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: updatedPagination,
+          )).thenAnswer(
         (_) async =>
             SearchMessagesResponse()..results = paginatedMessageResponseList,
       );
@@ -406,17 +360,17 @@ void main() {
         find.text([
           ...messageResponseList,
           ...paginatedMessageResponseList,
-        ].map((e) => '${e.channel.cid}-${e.message.id}').join(',')),
+        ].map((e) => '${e.channel?.cid}-${e.message.id}').join(',')),
         findsOneWidget,
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: updatedPagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: updatedPagination,
+          )).called(1);
     },
   );
 
@@ -426,8 +380,8 @@ void main() {
     (tester) async {
       const pagination = PaginationParams();
 
-      StateSetter _stateSetter;
-      int limit = pagination.limit;
+      StateSetter? _stateSetter;
+      var limit = pagination.limit;
 
       const messageSearchListCoreKey = Key('messageSearchListCore');
       const childWidgetKey = Key('childWidget');
@@ -438,7 +392,7 @@ void main() {
               key: childWidgetKey,
               child: Text(
                 messages
-                    .map((e) => '${e.channel.cid}-${e.message.id}')
+                    .map((e) => '${e.channel?.cid}-${e.message.id}')
                     .join(','),
               ),
             ),
@@ -446,18 +400,19 @@ void main() {
             emptyBuilder: (BuildContext context) => Offstage(),
             errorBuilder: (BuildContext context, Object error) => Offstage(),
             paginationParams: pagination.copyWith(limit: limit),
+            filters: {},
           );
 
       final mockClient = MockClient();
 
       final messageResponseList = _generateMessages();
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
@@ -483,32 +438,32 @@ void main() {
       expect(
         find.text(
           messageResponseList
-              .map((e) => '${e.channel.cid}-${e.message.id}')
+              .map((e) => '${e.channel?.cid}-${e.message.id}')
               .join(','),
         ),
         findsOneWidget,
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).called(1);
 
       // Rebuilding MessageSearchListCore with new pagination limit
-      _stateSetter(() => limit = 6);
+      _stateSetter?.call(() => limit = 6);
 
       final updatedMessageResponseList = _generateMessages(count: limit);
       final updatedPagination = pagination.copyWith(limit: limit);
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: updatedPagination,
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: updatedPagination,
+          )).thenAnswer(
         (_) async =>
             SearchMessagesResponse()..results = updatedMessageResponseList,
       );
@@ -518,18 +473,18 @@ void main() {
       expect(find.byKey(childWidgetKey), findsOneWidget);
       expect(
         find.text(updatedMessageResponseList
-            .map((e) => '${e.channel.cid}-${e.message.id}')
+            .map((e) => '${e.channel?.cid}-${e.message.id}')
             .join(',')),
         findsOneWidget,
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: updatedPagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: updatedPagination,
+          )).called(1);
     },
   );
 }
