@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_flutter_core/src/message_search_bloc.dart';
 import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
@@ -23,23 +23,11 @@ void main() {
             text: 'testTextData$index',
           )
           ..channel = ChannelModel(
-            cid: 'testCid',
+            cid: 'testCid:id',
           );
       },
     );
   }
-
-  test(
-    'should throw assertion error if child is null',
-    () async {
-      const messageSearchBlocKey = Key('messageSearchBloc');
-      final messageSearchBloc = () => MessageSearchBloc(
-            key: messageSearchBlocKey,
-            child: null,
-          );
-      expect(messageSearchBloc, throwsA(isA<AssertionError>()));
-    },
-  );
 
   testWidgets(
     'messageSearchBlocState.search() should throw if used where '
@@ -62,7 +50,7 @@ void main() {
       );
 
       try {
-        await usersBlocState.search();
+        await usersBlocState.search(filter: {});
       } catch (e) {
         expect(e, isInstanceOf<Exception>());
       }
@@ -93,30 +81,30 @@ void main() {
 
       final messageResponseList = _generateMessages();
 
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
-      messageSearchBlocState.search();
+      messageSearchBlocState.search(filter: {});
 
       await expectLater(
         messageSearchBlocState.messagesStream,
         emits(isSameMessageResponseListAs(messageResponseList)),
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
     },
   );
 
@@ -144,28 +132,28 @@ void main() {
       );
 
       const error = 'Error! Error! Error!';
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenThrow(error);
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenThrow(error);
 
-      messageSearchBlocState.search();
+      messageSearchBlocState.search(filter: {});
 
       await expectLater(
         messageSearchBlocState.messagesStream,
         emitsError(error),
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
     },
   );
 
@@ -196,47 +184,47 @@ void main() {
 
       final messageResponseList = _generateMessages();
 
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
-      messageSearchBlocState.search();
+      messageSearchBlocState.search(filter: {});
 
       await expectLater(
         messageSearchBlocState.messagesStream,
         emits(isSameMessageResponseListAs(messageResponseList)),
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
 
       final offset = messageResponseList.length;
       final paginatedMessageResponseList = _generateMessages(offset: offset);
       final pagination = PaginationParams(offset: offset);
 
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).thenAnswer(
         (_) async =>
             SearchMessagesResponse()..results = paginatedMessageResponseList,
       );
 
-      messageSearchBlocState.search(pagination: pagination);
+      messageSearchBlocState.search(pagination: pagination, filter: {});
 
       await Future.wait([
         expectLater(
@@ -251,13 +239,13 @@ void main() {
         ),
       ]);
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).called(1);
     },
   );
 
@@ -288,57 +276,57 @@ void main() {
 
       final messageResponseList = _generateMessages();
 
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).thenAnswer(
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).thenAnswer(
         (_) async => SearchMessagesResponse()..results = messageResponseList,
       );
 
-      messageSearchBlocState.search();
+      messageSearchBlocState.search(filter: {});
 
       await expectLater(
         messageSearchBlocState.messagesStream,
         emits(isSameMessageResponseListAs(messageResponseList)),
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: anyNamed('paginationParams'),
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: any(named: 'paginationParams'),
+          )).called(1);
 
       final offset = messageResponseList.length;
       final pagination = PaginationParams(offset: offset);
 
       const error = 'Error! Error! Error!';
-      when(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).thenThrow(error);
+      when(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).thenThrow(error);
 
-      messageSearchBlocState.search(pagination: pagination);
+      messageSearchBlocState.search(pagination: pagination, filter: {});
 
       await expectLater(
         messageSearchBlocState.queryMessagesLoading,
         emitsError(error),
       );
 
-      verify(mockClient.search(
-        any,
-        query: anyNamed('query'),
-        sort: anyNamed('sort'),
-        messageFilters: anyNamed('messageFilters'),
-        paginationParams: pagination,
-      )).called(1);
+      verify(() => mockClient.search(
+            any(),
+            query: any(named: 'query'),
+            sort: any(named: 'sort'),
+            messageFilters: any(named: 'messageFilters'),
+            paginationParams: pagination,
+          )).called(1);
     },
   );
 }
